feat: concurrency hardening, stress test suite, refresh OpenAPI + TS contracts
- Fix partial index WHERE clauses (text() not raw strings) in Hold and Booking models
- Add row locking (with_for_update) to confirm_hold and promote_waitlist
- Add uq_pending_holds index on (provider_id, start_time) WHERE pending
- Add uq_active_bookings index on (provider_id, start_time) WHERE not cancelled
- New test suites: test_concurrency, test_security_fuzzing, test_scheduling_edge_cases
- New test suites: test_audit_fixes, test_audit_resolution, test_multi_tenancy
- Document two concurrency vulnerabilities (SQLite xfail + runtime warning)
- Regenerate openapi.json (120 paths, 148 schemas)
- Regenerate contracts/types.ts via openapi-typescript 7.13.0
- Test result: 38 passed, 1 xpassed, 0 failed

+++ app/schemas/addon.py
@@ -1,6 +1,7 @@
 """Schemas for add‑on models."""
 
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict
@@ -10,7 +11,7 @@ class AddOnBase(BaseModel):
     service_id: int
     name: str
     description: Optional[str] = None
-    price: Optional[float] = None
+    price: Optional[Decimal] = None
     duration: int = 0
     active: bool = True
